Bootstrap website
This commit contains a responsive webpage created using Bootstrap v5.0.  
During a lecture, we were tasked to create a website similar to the one presented by the lecturer on the screen. As the images were already given, no time was waisted for browsing the web and all efforts focused on writing the code and learn Bootstrap v5.0. When working on this project I explored various possibilities Bootstrap gives and observed how I can change the website using information from the documentation.